Cambios en contratos de clientes y cotizaciones

+++ ERPAPI/Models/Facturacion/SalesOrder.cs
@@ -75,7 +75,7 @@ namespace ERPAPI.Models
         public DateTime OrderDate { get; set; }
         public DateTime ExpirationDate { get; set; }
 
-        public int? PlazoMeses { get; set; }
+        public decimal? PlazoMeses { get; set; }
         /// <summary>
         /// Precio Base del proucto a Almacenar
         /// </summary>
